+++ include/RenderRoutines.h
@@ -25,7 +25,7 @@ namespace RenderRoutines
     extern void drawString(std::string text, int x, int y);
     extern void drawString(std::string text, int x, int y, float scale);
     extern void drawString(std::string text, int x, int y, float scale, float r, float g, float b);
-    extern void isoBlit(int tex, int i, int j, int h1, int h2, int h3, int h4); //Used for drawing terrain mesh in 3D space isometrically
+    extern void isoBlit(int tex, double i, double j, int h1, int h2, int h3, int h4); //Used for drawing terrain mesh in 3D space isometrically
     extern void blit(int tex, int x, int y);
     extern void blit(int tex, int x1, int y1, int x2, int y2);
     extern void drawCuboid(Cuboid& c);
